requirements.txt
streamlit~=1.46.1
pandas~=2.3.1
together~=1.5.21
pydantic~=2.11.7
python-dotenv~=1.1.1
pymongo
pymongo[srv]